fixed accelerometer and recording timestamp mismatch; debugging new issue with actually using interal timestamps

+++ mease_lab_to_nwb/convert_syntalos/syntalosrecordingextractor.py
@@ -51,7 +51,7 @@ class SyntalosRecordingExtractor(MultiRecordingTimeExtractor):
         return self._timestamps[frame]
 
     def time_to_frame(self, time):
-        return np.searchsorted(self._timestamps, time).astype('int64')
+        return np.searchsorted(self._timestamps, time)
 
     def get_timestamps(self):
         return self._timestamps
